Split file "traversal.rs" into "bfs.rs" and "dfs.rs"
