+++ crates/algos/src/lib.rs
@@ -141,11 +141,12 @@
 //! ```
 
 pub mod afforest;
+pub mod bfs;
+pub mod dfs;
 pub mod dss;
 pub mod page_rank;
 pub mod prelude;
 pub mod sssp;
-pub mod traversal;
 pub mod triangle_count;
 pub mod utils;
 pub mod wcc;
